Improve expansion function for not argv append in only not finded params tok
