+++ source/expansion/expansion.c
@@ -6,52 +6,11 @@
 /*   By: migugar2 <migugar2@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/12 23:00:52 by migugar2          #+#    #+#             */
-/*   Updated: 2025/08/24 21:39:09 by migugar2         ###   ########.fr       */
+/*   Updated: 2025/08/25 20:28:17 by migugar2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-t_param	*expand_param(t_shell *shell, t_seg *seg)
-{
-	t_param	*res;
-	char	*key;
-
-	res = new_param();
-	if (res == NULL)
-		return (NULL);
-	if (seg->slice.len == 1 && seg->slice.begin[0] == '?')
-	{
-		res->value = ft_itoa(shell->last_status);
-		res->len = ft_strlen(res->value);
-		return (res);
-	}
-	key = malloc(sizeof(char) * (seg->slice.len + 1));
-	if (key == NULL)
-		return (free(res), NULL);
-	ft_strlcpy(key, seg->slice.begin, seg->slice.len + 1);
-	res->value = get_env_value(&shell->env_list, key);
-	free(key);
-	if (res->value == NULL)
-		res->value = "";
-	res->len = ft_strlen(res->value);
-	return (res);
-}
-
-size_t	consume_param(t_exp *exp, t_seg *seg, char *cur)
-{
-	t_param	*tmp;
-	size_t	len;
-
-	tmp = exp->head;
-	len = tmp->len;
-	ft_strlcpy(cur, tmp->value, tmp->len + 1);
-	exp->head = exp->head->next;
-	free_param(&tmp, seg);
-	if (exp->head == NULL)
-		exp->tail = NULL;
-	return (len);
-}
 
 char	*expand_tok(t_exp *exp, t_tok *tok)
 {
@@ -81,6 +40,20 @@ char	*expand_tok(t_exp *exp, t_tok *tok)
 	return (expanded);
 }
 
+static void	count_cur(t_exp	*exp, t_seg *cur)
+{
+	if (cur->type == SEG_TEXT)
+		exp->words += 1;
+	else if (cur->type == SEG_PARAM)
+	{
+		exp->paramc += 1;
+		if (exp->tail != NULL && exp->tail->value == NULL)
+			exp->paramnull += 1;
+	}
+	else if (cur->type == SEG_WILDCARD)
+		exp->wildcards += 1;
+}
+
 static int	get_info(t_shell *shell, t_exp *exp, t_tok *tok)
 {
 	t_seg	*cur;
@@ -101,10 +74,8 @@ static int	get_info(t_shell *shell, t_exp *exp, t_tok *tok)
 			exp->len += param->len;
 		}
 		else if (cur->type == SEG_WILDCARD)
-		{
-			exp->wildcards++;
 			exp->len++;
-		}
+		count_cur(exp, cur);
 		cur = cur->next;
 	}
 	return (0);
@@ -118,9 +89,14 @@ int	expansion(t_shell *shell, t_tok *tok, t_argv *argv)
 	exp.head = NULL;
 	exp.tail = NULL;
 	exp.len = 0;
+	exp.paramc = 0;
+	exp.paramnull = 0;
+	exp.words = 0;
 	exp.wildcards = 0;
 	if (get_info(shell, &exp, tok) == 1)
 		return (1);
+	if (exp.paramc > 0 && exp.paramnull == exp.paramc && exp.words == 0)
+		return (free_paramlst(&exp.head, &exp.tail, tok->seg_head), 0);
 	expanded = expand_tok(&exp, tok);
 	if (expanded == NULL)
 		return (free_paramlst(&exp.head, &exp.tail, tok->seg_head), 1);
